+++ src/settings/MyProfile.js
@@ -6,7 +6,7 @@ import { Settings } from '@folio/stripes/smart-components';
 import { stripesShape, TitleManager } from '@folio/stripes/core';
 
 import ChangePassword from './ChangePassword';
-import AppListSettings from './ApplListSettings';
+import AppListSettings from './AppOrderSettings';
 
 class MyProfile extends Component {
   static isLocalLogin(stripes) {
@@ -31,10 +31,10 @@ class MyProfile extends Component {
   createSettingsPages() {
     return [
       {
-        route: 'appList',
+        route: 'app-order',
         label: <FormattedMessage id="ui-myprofile.settings.appNavOrder.label" />,
         component: AppListSettings,
-        perm: 'ui-myprofile.settings.main-nav-order',
+        perm: 'ui-myprofile.settings.main-nav-order.manage',
       }
     ]
       .concat(
